+++ lib/hubbado/log.rb
@@ -3,6 +3,7 @@ require "hubbado/log/tags"
 require "hubbado/log/configuration"
 require "hubbado/log/log"
 require "hubbado/log/logger"
+require "hubbado/log/logger/substitute"
 require "hubbado/log/log_handler"
 
 module Hubbado